+++ packages/youtube_player_iframe/example/lib/main.dart
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'dart:developer';
+
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:youtube_player_iframe/youtube_player_iframe.dart';
@@ -53,24 +55,28 @@ class _YoutubeAppDemoState extends State<YoutubeAppDemo> {
         showFullscreenButton: true,
         autoPlay: true,
       ),
-    );
-
-    _controller.loadPlaylist(
-      list: [
-        'tcodrIK2P_I',
-        'nPt8bK2gbaU',
-        'K18cpp_-gP8',
-        'iLnmTe5Q2Qw',
-        '_WoCV4c6XOE',
-        'KmzdUe0RSJo',
-        '6jZDSSZZxjQ',
-        'p2lYr3vM_1w',
-        '7QUtEmBT_-w',
-        '34_PXCzGw1M',
-      ],
-      listType: ListType.playlist,
-      startSeconds: 136,
-    );
+    )
+      ..onInit = () {
+        _controller.loadPlaylist(
+          list: [
+            'tcodrIK2P_I',
+            'nPt8bK2gbaU',
+            'K18cpp_-gP8',
+            'iLnmTe5Q2Qw',
+            '_WoCV4c6XOE',
+            'KmzdUe0RSJo',
+            '6jZDSSZZxjQ',
+            'p2lYr3vM_1w',
+            '7QUtEmBT_-w',
+            '34_PXCzGw1M',
+          ],
+          listType: ListType.playlist,
+          startSeconds: 136,
+        );
+      }
+      ..onFullscreenChange = (isFullScreen) {
+        log('${isFullScreen ? 'Entered' : 'Exited'} Fullscreen.');
+      };
   }
 
   @override